+++ plugins/module_utils/managedcluster_addons/iam_policy_controller.py
@@ -1,0 +1,40 @@
+from __future__ import (absolute_import, division, print_function)
+
+__metaclass__ = type
+
+from ansible.module_utils.basic import AnsibleModule
+from .addon_base import addon_base
+
+
+# subclass
+class iam_policy_controller(addon_base):
+    def __init__(self, module: AnsibleModule, enabled, hub_client, managed_cluster_name, addon_name, wait=False, timeout=60):
+        super().__init__(module, enabled, hub_client,
+                         managed_cluster_name, addon_name, wait, timeout)
+
+    def check_feature(self):
+        self.check_cluster_management_addon_feature(
+            self.module,
+            self.hub_client,
+            self.addon_name
+        )
+
+    def enable_addon(self):
+        return self.enable_klusterlet_addon(
+            self.module,
+            self.hub_client,
+            self.managed_cluster_name,
+            self.addon_name,
+            self.wait,
+            self.timeout
+        )
+
+    def disable_addon(self):
+        return self.disable_klusterlet_addon(
+            self.module,
+            self.hub_client,
+            self.managed_cluster_name,
+            self.addon_name,
+            self.wait,
+            self.timeout
+        )
