+++ plugins/modules/managed_serviceaccount.py
@@ -109,9 +109,7 @@ import traceback
 
 from ansible.module_utils.basic import AnsibleModule, env_fallback, missing_required_lib
 from ansible_collections.ocmplus.cm.plugins.module_utils.import_utils import get_managed_cluster
-from ansible_collections.ocmplus.cm.plugins.module_utils.addon_utils import (
-    check_addon_available,
-)
+from ansible_collections.ocmplus.cm.plugins.module_utils.addon_utils import check_addon_available
 
 IMP_ERR = {}
 try:
