+++ tests/integration/targets/multicluster_management/tasks/multicluster-management-playbook.yml
@@ -3,12 +3,12 @@
   connection: local
   tasks:
   - name: "Install cluster-proxy addon for {{ hostvars[inventory_hostname].cluster_name }}"
-    ocmplus.cm.managed_cluster_addon:
+    ocmplus.cm.managedcluster_addon:
       hub_kubeconfig: "{{ hostvars['local-cluster'].kubeconfig }}"
       managed_cluster: "{{ hostvars[inventory_hostname].cluster_name }}"
       addon_name: "cluster-proxy"
       wait: True
-      timeout: 120
+      timeout: 180
   - name: "Get ClusterProxy URL for {{ hostvars[inventory_hostname].cluster_name }}"
     ocmplus.cm.cluster_proxy:
       hub_kubeconfig: "{{ hostvars['local-cluster'].kubeconfig }}"
@@ -20,7 +20,7 @@
     debug:
       msg: "{{ cluster_proxy_url.cluster_url }}"
   - name: "Install managed-serviceaccount addon for {{ hostvars[inventory_hostname].cluster_name }}"
-    ocmplus.cm.managed_cluster_addon:
+    ocmplus.cm.managedcluster_addon:
       hub_kubeconfig: "{{ hostvars['local-cluster'].kubeconfig }}"
       managed_cluster: "{{ hostvars[inventory_hostname].cluster_name }}"
       addon_name: "managed-serviceaccount"
